# source/src/model/checklistItem.model.go
package model

import (
	"todolist-api/src/util/umongo"

	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/bson/primitive"
)

// Generated by https://quicktype.io

type ChecklistItem struct {
	MetadataWithID `bson:",inline"`

	ItemName    string `json:"Itemname" bson:"Itemname"`
	Description string `json:"description" bson:"description"`
	CheckListId string `json:"checklist_id" bson:"checklist_id"`
	Status      bool   `json:"status" bson:"status"`
}

type ChecklistItem_View struct {
	ChecklistItem `bson:",inline"`
}

type ChecklistItem_Search struct {
	//? Regex
	Search      string `json:"search"`
	CheckListId string `json:"checkListId"`

	umongo.Request
}

type CheklistItem_Update_status struct {
	Status bool `json:"status" bson:"status"`
}

type CheklistItem_Update_name struct {
	Itemname string `json:"Itemname" bson:"Itemname"`
}

func (o *ChecklistItem_Search) HandleFilter(listFilterAnd *[]bson.M) {
	if search := o.Search; search != "" {
		*listFilterAnd = append(*listFilterAnd, bson.M{"name": primitive.Regex{Pattern: search, Options: "i"}})
	}

	if CheckListId := o.CheckListId; CheckListId != "" {
		*listFilterAnd = append(*listFilterAnd, bson.M{"checklist_id": CheckListId})
	}
}